Improve reordering experience
