+++ podcasts/Podcasts/All Podcasts/PodcastListViewController+Edit.swift
@@ -12,7 +12,17 @@ extension PodcastListViewController {
         }
     }
 
-    @objc func doneEditingTapped() {
+    @objc func saveEditingTapped() {
+        saveSortOrder()
+        Analytics.track(.podcastsListReordered)
+        setEditingOrder(false)
+    }
+
+    @objc func cancelEditingTapped() {
+        if let snapshot = orderBeforeEditing, snapshot != gridItems {
+            gridItems = snapshot
+            podcastsCollectionView.reloadData()
+        }
         setEditingOrder(false)
     }
 
@@ -35,10 +45,12 @@ extension PodcastListViewController {
         savedLeftBarButtonItem = navigationItem.leftBarButtonItem
         savedRightBarButtonItem = customRightBtn
 
-        navigationItem.leftBarButtonItem = nil
-        customRightBtn = UIBarButtonItem(barButtonSystemItem: .done, target: self, action: #selector(doneEditingTapped))
+        navigationItem.leftBarButtonItem = UIBarButtonItem(barButtonSystemItem: .cancel, target: self, action: #selector(cancelEditingTapped))
+        customRightBtn = UIBarButtonItem(barButtonSystemItem: .save, target: self, action: #selector(saveEditingTapped))
 
+        orderBeforeEditing = gridItems
         podcastsCollectionView.dragInteractionEnabled = true
+        setTabBarHidden(true, animated: true)
 
         for cell in podcastsCollectionView.visibleCells {
             applyEditingTreatment(to: cell)
@@ -46,7 +58,9 @@ extension PodcastListViewController {
     }
 
     private func exitEditMode() {
+        orderBeforeEditing = nil
         podcastsCollectionView.dragInteractionEnabled = false
+        setTabBarHidden(false, animated: true)
 
         navigationItem.leftBarButtonItem = savedLeftBarButtonItem
         customRightBtn = savedRightBarButtonItem
@@ -58,6 +72,11 @@ extension PodcastListViewController {
         }
     }
 
+    private func setTabBarHidden(_ hidden: Bool, animated: Bool) {
+        guard #available(iOS 18.0, *), let tabBarController else { return }
+        tabBarController.setTabBarHidden(hidden, animated: animated)
+    }
+
     // MARK: Wiggle (grid)
 
     private static let wiggleAnimationKey = "podcasts.editingWiggle"
@@ -65,8 +84,8 @@ extension PodcastListViewController {
     private func startWiggle(on cell: UICollectionViewCell) {
         guard cell.layer.animation(forKey: Self.wiggleAnimationKey) == nil else { return }
         let animation = CAKeyframeAnimation(keyPath: "transform.rotation.z")
-        animation.values = [-0.022, 0.022, -0.022]
-        animation.duration = 0.16
+        animation.values = [-0.012, 0.012, -0.012]
+        animation.duration = 0.28
         animation.repeatCount = .infinity
         animation.timeOffset = .random(in: 0...animation.duration) // stagger so cells don't move in unison
         cell.layer.add(animation, forKey: Self.wiggleAnimationKey)
@@ -98,5 +117,52 @@ extension PodcastListViewController {
 
     private func removeReorderHandle(from cell: UICollectionViewCell) {
         cell.contentView.viewWithTag(Self.reorderHandleTag)?.removeFromSuperview()
+    }
+}
+
+extension PodcastListViewController: UICollectionViewDragDelegate, UICollectionViewDropDelegate {
+    // MARK: - UICollectionViewDragDelegate
+
+    func collectionView(_ collectionView: UICollectionView, itemsForBeginning session: UIDragSession, at indexPath: IndexPath) -> [UIDragItem] {
+        guard isEditingOrder, let item = itemAt(indexPath: indexPath), !item.isEmpty else {
+            return []
+        }
+        let provider = NSItemProvider(object: (item.podcast?.uuid ?? item.folder?.uuid ?? "") as NSString)
+        let dragItem = UIDragItem(itemProvider: provider)
+        dragItem.localObject = item
+        return [dragItem]
+    }
+
+    func collectionView(_ collectionView: UICollectionView, dragSessionIsRestrictedToDraggingApplication session: UIDragSession) -> Bool {
+        true
+    }
+
+    // MARK: - UICollectionViewDropDelegate
+
+    func collectionView(_ collectionView: UICollectionView, canHandle session: UIDropSession) -> Bool {
+        session.localDragSession != nil
+    }
+
+    func collectionView(_ collectionView: UICollectionView, dropSessionDidUpdate session: UIDropSession, withDestinationIndexPath destinationIndexPath: IndexPath?) -> UICollectionViewDropProposal {
+        guard session.localDragSession != nil else {
+            return UICollectionViewDropProposal(operation: .forbidden)
+        }
+        return UICollectionViewDropProposal(operation: .move, intent: .insertAtDestinationIndexPath)
+    }
+
+    func collectionView(_ collectionView: UICollectionView, performDropWith coordinator: UICollectionViewDropCoordinator) {
+        guard let dropItem = coordinator.items.first,
+              let sourceIndexPath = dropItem.sourceIndexPath else {
+            return
+        }
+        let destinationIndexPath = coordinator.destinationIndexPath
+            ?? IndexPath(item: max(0, gridItems.count - 1), section: 0)
+
+        collectionView.performBatchUpdates {
+            let moved = gridItems.remove(at: sourceIndexPath.item)
+            gridItems.insert(moved, at: destinationIndexPath.item)
+            collectionView.moveItem(at: sourceIndexPath, to: destinationIndexPath)
+        }
+        coordinator.drop(dropItem.dragItem, toItemAt: destinationIndexPath)
     }
 }
